Add spec for sign out
