+++ spec/support/sign_in.rb
@@ -1,0 +1,8 @@
+def sign_in(email, password)
+  visit new_user_session_path
+
+  fill_in "Email", with: email
+  fill_in "Password", with: password
+
+  click_button "Log in"
+end
