+++ spec/support/sign_out.rb
@@ -1,0 +1,10 @@
+def sign_out(email, password)
+  visit new_user_session_path
+
+  fill_in "Email", with: email
+  fill_in "Password", with: password
+
+  click_button "Log in"
+  click_link "Sign out"
+  page.accept_alert
+end
